Fixed degree and subspecialty dynamic fetching, now working on fixing connection requests
